Added an admin page for POST and GET

+++ src/main/java/coursework/ecomarket/repositories/CartRepo.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 
 import coursework.ecomarket.entities.Carts;
 
-public interface CartRepo extends CrudRepository<Carts, Long>{
+public interface CartRepo extends CrudRepository<Carts, Integer>{
 }
