+++ src/main/java/coursework/ecomarket/repositories/ClientRepo.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import coursework.ecomarket.entities.Client;
 
-public interface ClientRepo extends CrudRepository<Client, Long> {
+public interface ClientRepo extends CrudRepository<Client, Integer> {
     Client findByUsername(String username);
     Client findById(int id);
 }
