+++ src/main/java/coursework/ecomarket/repositories/OrderRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import coursework.ecomarket.entities.Order;
 
-public interface OrderRepo extends CrudRepository<Order, Long>{ 
+public interface OrderRepo extends CrudRepository<Order, Integer>{
     public Order getById(int id);
 }
